infinite loop of byte

+++ src/Main.java
@@ -3,7 +3,7 @@ import java.util.*;
 public class Main {
     public static void main(String[] args)
     {
-        int i=1;
+        byte i=1;
    while (true)
          {
             System.out.println(i);
